Interpolator speedup/slowdown algorithm version 2

+++ modules/network_synchronizer/net_utilities.cpp
@@ -34,6 +34,8 @@
 
 #include "net_utilities.h"
 
+#include "core/typedefs.h"
+
 NetworkTracer::NetworkTracer(int p_packets_to_track) :
 		id(0) {
 	reset(p_packets_to_track);
@@ -47,7 +49,7 @@ void NetworkTracer::reset(int p_packets_to_track) {
 	}
 
 	// Let's pretend that the connection is good.
-	for (size_t i = 0; i < flags.size(); i += 1) {
+	for (uint32_t i = 0; i < flags.size(); i += 1) {
 		flags[i] = true;
 	}
 }
